# api/openapi.yaml
openapi: 3.1.0
info:
  title: (S)AGE API
  version: 11.17.0
  description: |-
    (Sovereign) Agent Governed Experience — network REST API for memory
    submission, validation, and query.

    This machine-readable contract intentionally excludes the same-machine
    human CEREBRUM control plane under `/v1/dashboard/**`. Those routes are
    loopback-only management APIs, not remotely callable network APIs. Their
    authoritative contract is the “App-v23 CEREBRUM access-control routes”
    section of `docs/reference/rest-api.md`.
  contact:
    name: Dhillon Andrew Kannabhiran
    email: dhillon@levelupctf.com
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development

# ---------------------------------------------------------------------------
# Security — Ed25519 signature authentication via headers
# ---------------------------------------------------------------------------
security:
  - ed25519Auth: []
components:
...
  securitySchemes:
    ed25519Auth:
      type: apiKey
      in: header
      name: X-Agent-ID
      description: >-
        Ed25519 authentication is conveyed through three headers sent together:
        **X-Agent-ID** (hex-encoded public key), **X-Signature** (Ed25519
        signature of the canonical request payload), and **X-Timestamp** (unix
        epoch seconds). All three must be present on authenticated endpoints.
        Current clients should also send **X-Nonce**. After app-v17 activation,
        consensus binds delegated REST proofs to the exact canonical request,
        enforces the timestamp against block time, and consumes each proof once.

  # -----------------------------------------------------------------------
  # Parameters
  # -----------------------------------------------------------------------
...
  schemas:
...
    RegisteredAgent:
      type: object
      properties:
        agent_id:
          type: string
          description: Hex-encoded Ed25519 public key.
        name:
          type: string
        role:
          type: string
        boot_bio:
          type: string
        provider:
          type: string
        status:
          type: string
        clearance:
          type: integer
        on_chain_height:
          type: integer
          format: int64
          description: Block height where registered on-chain.
        visible_agents:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details for HTTP APIs.
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          description: URI reference identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        reason_code:
          type: string
          description: Stable sanitized machine-readable denial reason, when available.
        remedy:
          type: string
          description: Exact operator action that can make a later request valid.
        retryable:
          type: boolean
          description: Whether automatically repeating the same request is safe.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication headers.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    Forbidden:
      description: >-
        Caller authenticated but lacks authorisation, or the request was
        rejected during on-chain finalisation (e.g. quorum-gated tx without
        an executed proposal). The `detail` field carries the upstream
        rejection reason.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    InternalError:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
paths:
...
  /v1/agents:
    get:
      operationId: listRegisteredAgents
      summary: List agents visible to the authenticated caller.
      description: >-
        Requires an Ed25519-signed request. After app-v23 activation the
        response contains only active ordinary agents within the caller's
        enforced discovery boundary; pending, inactive, Root, and inconsistent
        identities are omitted.
      tags:
        - Agent
      responses:
        "200":
          description: Caller-scoped list of active ordinary agents.
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/RegisteredAgent"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...